mid refactor - seperated to abstract handler, resource and module
